Modernization

+++ UnshieldSharp/FileDescriptor.cs
@@ -5,9 +5,9 @@
         public uint NameOffset;
         public uint DirectoryIndex;
         public ushort Flags;
-        public uint ExpandedSize;
-        public uint CompressedSize;
-        public uint DataOffset;
+        public ulong ExpandedSize;
+        public ulong CompressedSize;
+        public ulong DataOffset;
         public byte[] Md5 = new byte[16];
         public ushort Volume;
         public uint LinkPrevious;
